[VIRT] Move fixtures used exclusively by tests/virt
Move fixtures that are used exclusively by tests/virt from the main tests/conftest.py to tests/virt/conftest.py for better organization and proximity to their usage.
